Move legacy XML parser for `pw.x` to separate file
The XML parser for the old format, deprecated since Quantum ESPRESSO v6.4
has been moved to its own file for clarity.
