+++ aiida_quantumespresso/parsers/pw.py
@@ -87,107 +87,7 @@ class PwParser(Parser):
         all_symmetries = parser_options.get('all_symmetries', False)
 
         if not all_symmetries and 'cell' in structure_data:
-
-            # In the standard output, each symmetry operation print two rotation matrices:
-            #
-            # * S_cryst^T: matrix in crystal coordinates, transposed
-            # * S_cart: matrix in cartesian coordinates,
-            #
-            # The XML files only print one matrix:
-            #
-            # * S_cryst: matrix in crystal coordinates
-            #
-            # The raw parsed symmetry information from the XML is large and will load the database heavily if stored as
-            # is for each calculation. Instead, we will map these dictionaries onto a static dictionary of rotation
-            # matrices generated by the _get_qe_symmetry_list static method. This dictionary will return the rotation
-            # matrices in cartesian coordinates, i.e. S_cart. In order to compare the raw matrices from the XML to these
-            # static matrices we have to convert S_cryst into S_cart. We derive here how that is done:
-            #
-            #   S_cryst * v_cryst = v_cryst'
-            #
-            # where v_cryst' is the rotated vector v_cryst under S_cryst
-            # We define `cell` where cell vectors are rows. Converting a vector from crystal to cartesian
-            # coordinates is defined as:
-            #
-            #   cell^T * v_cryst = v_cart
-            #
-            # The inverse of this operation is defined as
-            #
-            #   v_cryst = cell^Tinv * v_cart
-            #
-            # Replacing the last equation into the first we find:
-            #
-            #   S_cryst * cell^Tinv * v_cart = cell^Tinv * v_cart'
-            #
-            # Multiply on the left with cell^T gives:
-            #
-            #   cell^T * S_cryst * cell^Tinv * v_cart = v_cart'
-            #
-            # which can be rewritten as:
-            #
-            #   S_cart * v_cart = v_cart'
-            #
-            # where:
-            #
-            #   S_cart = cell^T * S_cryst * cell^Tinv
-            #
-            # We compute here the transpose and its inverse of the structure cell basis, which is needed to transform
-            # the parsed rotation matrices, which are in crystal coordinates, to cartesian coordinates, which are the
-            # matrices that are returned by the _get_qe_symmetry_list staticmethod
-            cell = structure_data['cell']['lattice_vectors']
-            cell_T = numpy.transpose(cell)
-            cell_Tinv = numpy.linalg.inv(cell_T)
-            possible_symmetries = self._get_qe_symmetry_list()
-
-            for symmetry_type in ['symmetries','lattice_symmetries']:  # crystal vs. lattice symmetries
-                if symmetry_type in list(out_dict.keys()):
-                    try:
-                        old_symmetries = out_dict[symmetry_type]
-                        new_symmetries = []
-                        for this_sym in old_symmetries:
-                            name = this_sym['name'].strip()
-                            for i, this in enumerate(possible_symmetries):
-                                # Since we do an exact comparison we strip the string name from whitespace
-                                # and as soon as it is matched, we break to prevent it from matching another
-                                if name == this['name'].strip():
-                                    index = i
-                                    break
-                            else:
-                                index = None
-                                self.logger.error('Symmetry {} not found'.format(name))
-
-                            new_dict = {}
-                            if index is not None:
-                                # The raw parsed rotation matrix is in crystal coordinates, whereas the mapped rotation
-                                # in possible_symmetries is in cartesian coordinates. To allow them to be compared
-                                # to make sure we matched the correct rotation symmetry, we first convert the parsed matrix
-                                # to cartesian coordinates. For explanation of the method, see comment above.
-                                rotation_cryst = this_sym['rotation']
-                                rotation_cart_new = possible_symmetries[index]['matrix']
-                                rotation_cart_old = numpy.dot(cell_T, numpy.dot(rotation_cryst, cell_Tinv))
-
-                                inversion = possible_symmetries[index]['inversion']
-                                if not are_matrices_equal(rotation_cart_old, rotation_cart_new, swap_sign_matrix_b=inversion):
-                                    self.logger.error('Mapped rotation matrix {} does not match the original rotation {}'
-                                        .format(rotation_cart_new, rotation_cart_old))
-                                    new_dict['all_symmetries'] = this_sym
-                                else:
-                                    # Note: here I lose the information about equivalent ions and fractional_translation
-                                    # since I don't copy them to new_dict (but they can be reconstructed).
-                                    new_dict['t_rev'] = this_sym['t_rev']
-                                    new_dict['symmetry_number'] = index
-                            else:
-                                new_dict['all_symmetries'] = this_sym
-
-                            new_symmetries.append(new_dict)
-
-                        out_dict[symmetry_type] = new_symmetries  # and overwrite the old one
-                    except KeyError:
-                        self.logger.warning("key '{}' is not present in raw output dictionary".format(symmetry_type))
-                else:
-                    # backwards-compatiblity: 'lattice_symmetries' is not created in older versions of the parser
-                    if symmetry_type != 'lattice_symmetries':
-                        self.logger.warning("key '{}' is not present in raw output dictionary".format(symmetry_type))
+            self.reduce_symmetries(out_dict, structure_data)
 
         # I eventually save the new structure. structure_data is unnecessary after this
         input_structure = self.node.get_incoming(link_label_filter='structure').one().node
@@ -196,16 +96,16 @@ class PwParser(Parser):
         if type_calc in ['relax', 'vc-relax', 'md', 'vc-md']:
             if 'cell' in list(structure_data.keys()):
                 output_structure = convert_qe2aiida_structure(structure_data, input_structure=input_structure)
-                self.out(self.get_linkname_outstructure(), output_structure)
+                self.out('output_structure', output_structure)
 
         k_points_list = trajectory_data.pop('k_points', None)
         k_points_weights_list = trajectory_data.pop('k_points_weights', None)
 
         if k_points_list is not None:
 
-            # Build the kpoints object
             if out_dict.pop('k_points_units') != '1 / angstrom':
-                raise QEOutputParsingError('Error in kpoints units (should be cartesian)')
+                self.logger.error('Error in kpoints units (should be cartesian)')
+                return self.exit_codes.ERROR_INVALID_KPOINT_UNITS
 
             kpoints_from_output = orm.KpointsData()
             kpoints_from_output.set_cell_from_structure(output_structure)
@@ -216,7 +116,7 @@ class PwParser(Parser):
                 try:
                     kpoints_from_input.get_kpoints()
                 except AttributeError:
-                    self.out(self.get_linkname_out_kpoints(), kpoints_from_output)
+                    self.out('output_kpoints', kpoints_from_output)
 
             # Converting bands into a BandsData object (including the kpoints)
             if bands_data:
@@ -291,14 +191,14 @@ class PwParser(Parser):
 
                 for x in six.iteritems(trajectory_data):
                     traj.set_array(x[0], numpy.array(x[1]))
-                self.out(self.get_linkname_outtrajectory(), traj)
+                self.out('output_trajectory', traj)
 
             except KeyError:
                 # forces, atomic charges and atomic mag. moments, in scf calculation (when outputed)
                 arraydata = orm.ArrayData()
                 for x in six.iteritems(trajectory_data):
                     arraydata.set_array(x[0], numpy.array(x[1]))
-                self.out(self.get_linkname_outarray(), arraydata)
+                self.out('output_array', arraydata)
 
         return ExitCode(0)
 
@@ -309,62 +209,137 @@ class PwParser(Parser):
         """
         return 'parser_options'
 
-    def get_linkname_outstructure(self):
-        """
-        Returns the name of the link to the output_structure
-        Node exists if positions or cell changed.
-        """
-        return 'output_structure'
-
-    def get_linkname_outtrajectory(self):
-        """
-        Returns the name of the link to the output_trajectory.
-        Node exists in case of calculation='md', 'vc-md', 'relax', 'vc-relax'
-        """
-        return 'output_trajectory'
-
-    def get_linkname_outarray(self):
-        """
-        Returns the name of the link to the output_array
-        Node may exist in case of calculation='scf'
-        """
-        return 'output_array'
-
-    def get_linkname_out_kpoints(self):
-        """
-        Returns the name of the link to the output_kpoints
-        Node exists if cell has changed and no bands are stored.
-        """
-        return 'output_kpoints'
-
     def get_extended_symmetries(self):
-        """
-        Return the extended dictionary of symmetries.
-        """
+        """Return the extended dictionary of symmetries based on reduced symmetries stored in output parameters."""
         possible_symmetries = self._get_qe_symmetry_list()
-        data = self.node.get_outgoing(node_class=orm.Dict)
-        all_data = [i.node for i in data if i.link_label == self.get_linkname_outparams()]
-        if len(all_data) > 1:
-            raise exceptions.UniquenessError('More than one output parameterdata found.')
-        elif not all_data:
-            return []
-        else:
-            compact_list = all_data[0].get_dict()['symmetries']  # rimetti lo zero
-            new_list = []
-            # copy what wasn't compact
-            for element in compact_list:
-                new_dict = {}
-                for keys in ['t_rev', 'equivalent_ions', 'fractional_translation']:
-                    try:
-                        new_dict[keys] = element[keys]
-                    except KeyError:
-                        pass
-                # expand the rest
-                new_dict['name'] = possible_symmetries[element['symmetry_number']]['name']
-                new_dict['rotation'] = possible_symmetries[element['symmetry_number']]['matrix']
-                new_dict['inversion'] = possible_symmetries[element['symmetry_number']]['inversion']
-                new_list.append(new_dict)
-            return new_list
+        parameters = self.node.get_outgoing(node_class=orm.Dict).get_node_by_label('output_parameters')
+
+        symmetries_extended = []
+        symmetries_reduced = parameters.get_dict()['symmetries']  # rimetti lo zero
+
+        for element in symmetries_reduced:
+
+            symmetry = {}
+
+            for keys in ['t_rev', 'equivalent_ions', 'fractional_translation']:
+                try:
+                    symmetry[keys] = element[keys]
+                except KeyError:
+                    pass
+
+            # expand the rest
+            symmetry['name'] = possible_symmetries[element['symmetry_number']]['name']
+            symmetry['rotation'] = possible_symmetries[element['symmetry_number']]['matrix']
+            symmetry['inversion'] = possible_symmetries[element['symmetry_number']]['inversion']
+
+            symmetries_extended.append(symmetry)
+
+        return symmetries_extended
+
+    def reduce_symmetries(self, out_dict, structure_data):
+        """Reduce the symmetry information parsed from the output to save space.
+
+        In the standard output, each symmetry operation print two rotation matrices:
+
+            * S_cryst^T: matrix in crystal coordinates, transposed
+            * S_cart: matrix in cartesian coordinates,
+
+        The XML files only print one matrix:
+
+            * S_cryst: matrix in crystal coordinates
+
+        The raw parsed symmetry information from the XML is large and will load the database heavily if stored as
+        is for each calculation. Instead, we will map these dictionaries onto a static dictionary of rotation
+        matrices generated by the _get_qe_symmetry_list static method. This dictionary will return the rotation
+        matrices in cartesian coordinates, i.e. S_cart. In order to compare the raw matrices from the XML to these
+        static matrices we have to convert S_cryst into S_cart. We derive here how that is done:
+
+            S_cryst * v_cryst = v_cryst'
+
+        where v_cryst' is the rotated vector v_cryst under S_cryst
+        We define `cell` where cell vectors are rows. Converting a vector from crystal to cartesian
+        coordinates is defined as:
+
+            cell^T * v_cryst = v_cart
+
+        The inverse of this operation is defined as
+
+            v_cryst = cell^Tinv * v_cart
+
+        Replacing the last equation into the first we find:
+
+            S_cryst * cell^Tinv * v_cart = cell^Tinv * v_cart'
+
+        Multiply on the left with cell^T gives:
+
+            cell^T * S_cryst * cell^Tinv * v_cart = v_cart'
+
+        which can be rewritten as:
+
+            S_cart * v_cart = v_cart'
+
+        where:
+
+            S_cart = cell^T * S_cryst * cell^Tinv
+
+        We compute here the transpose and its inverse of the structure cell basis, which is needed to transform
+        the parsed rotation matrices, which are in crystal coordinates, to cartesian coordinates, which are the
+        matrices that are returned by the _get_qe_symmetry_list staticmethod
+        """
+        cell = structure_data['cell']['lattice_vectors']
+        cell_T = numpy.transpose(cell)
+        cell_Tinv = numpy.linalg.inv(cell_T)
+        possible_symmetries = self._get_qe_symmetry_list()
+
+        for symmetry_type in ['symmetries', 'lattice_symmetries']:  # crystal vs. lattice symmetries
+            if symmetry_type in list(out_dict.keys()):
+                try:
+                    old_symmetries = out_dict[symmetry_type]
+                    new_symmetries = []
+                    for this_sym in old_symmetries:
+                        name = this_sym['name'].strip()
+                        for i, this in enumerate(possible_symmetries):
+                            # Since we do an exact comparison we strip the string name from whitespace
+                            # and as soon as it is matched, we break to prevent it from matching another
+                            if name == this['name'].strip():
+                                index = i
+                                break
+                        else:
+                            index = None
+                            self.logger.error('Symmetry {} not found'.format(name))
+
+                        new_dict = {}
+                        if index is not None:
+                            # The raw parsed rotation matrix is in crystal coordinates, whereas the mapped rotation
+                            # in possible_symmetries is in cartesian coordinates. To allow them to be compared
+                            # to make sure we matched the correct rotation symmetry, we first convert the parsed matrix
+                            # to cartesian coordinates. For explanation of the method, see comment above.
+                            rotation_cryst = this_sym['rotation']
+                            rotation_cart_new = possible_symmetries[index]['matrix']
+                            rotation_cart_old = numpy.dot(cell_T, numpy.dot(rotation_cryst, cell_Tinv))
+
+                            inversion = possible_symmetries[index]['inversion']
+                            if not are_matrices_equal(rotation_cart_old, rotation_cart_new, swap_sign_matrix_b=inversion):
+                                self.logger.error('Mapped rotation matrix {} does not match the original rotation {}'
+                                    .format(rotation_cart_new, rotation_cart_old))
+                                new_dict['all_symmetries'] = this_sym
+                            else:
+                                # Note: here I lose the information about equivalent ions and fractional_translation
+                                # since I don't copy them to new_dict (but they can be reconstructed).
+                                new_dict['t_rev'] = this_sym['t_rev']
+                                new_dict['symmetry_number'] = index
+                        else:
+                            new_dict['all_symmetries'] = this_sym
+
+                        new_symmetries.append(new_dict)
+
+                    out_dict[symmetry_type] = new_symmetries  # and overwrite the old one
+                except KeyError:
+                    self.logger.warning("key '{}' is not present in raw output dictionary".format(symmetry_type))
+            else:
+                # backwards-compatiblity: 'lattice_symmetries' is not created in older versions of the parser
+                if symmetry_type != 'lattice_symmetries':
+                    self.logger.warning("key '{}' is not present in raw output dictionary".format(symmetry_type))
 
     @staticmethod
     def _get_qe_symmetry_list():
